+++ src/main/java/State/State.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 public class State implements Cloneable {
     // 0 = empty, 1 = black queen, 2 = white queen, 3 = arrow
     private byte[][] board;
-    private int[][] blackQueens;
-    private int[][] whiteQueens;
+    private final int[][] blackQueens;
+    private final int[][] whiteQueens;
 
     public final int BOARD_SIZE = 10;
     public static final int WHITE_QUEEN = 2;
@@ -39,9 +39,30 @@ public class State implements Cloneable {
         try {
             State cloned = (State) state.clone();
             board = cloned.board;
-            board[action.getNewPos().get(0)][action.getNewPos().get(1)] = board[action.getOldPos().get(0)][action.getOldPos().get(1)];
+            blackQueens = cloned.blackQueens;
+            whiteQueens = cloned.whiteQueens;
+
+            byte movingQueen = board[action.getOldPos().get(0)][action.getOldPos().get(1)];
+            board[action.getNewPos().get(0)][action.getNewPos().get(1)] = movingQueen;
             board[action.getOldPos().get(0)][action.getOldPos().get(1)] = 0;
             board[action.getArrowPos().get(0)][action.getArrowPos().get(1)] = 3;
+
+            // Update the queen that moved
+            if (movingQueen == BLACK_QUEEN) {
+                for (int i = 0; i < blackQueens.length; i++) {
+                    if (blackQueens[i][0] == action.getOldPos().get(0) && blackQueens[i][1] == action.getOldPos().get(1)) {
+                        blackQueens[i] = new int[]{action.getNewPos().get(0), action.getNewPos().get(1)};
+                        break;
+                    }
+                }
+            } else {
+                for (int i = 0; i < whiteQueens.length; i++) {
+                    if (whiteQueens[i][0] == action.getOldPos().get(0) && whiteQueens[i][1] == action.getOldPos().get(1)) {
+                        whiteQueens[i] = new int[]{action.getNewPos().get(0), action.getNewPos().get(1)};
+                        break;
+                    }
+                }
+            }
 
         } catch (CloneNotSupportedException e) {
             throw new RuntimeException(e);
